did some pylint changes and renamed NetworkCommander folder to networkcommander
